% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{multisegsteg-package}
\alias{multisegsteg-package}
\alias{multisegsteg}
\title{multisegsteg: Regression Models with Multiple Break-Points Estimation}
\description{
Implements methods for fitting and visualizing segmented regression models
with multiple thresholds, including multi-segmented regression (continuous
at thresholds) and multi-stegmented regression (allowing discontinuities).
}
\details{
The package provides tools for:
\itemize{
\item Estimation of multiple thresholds in segmented and stegmented regression models
\item Support for both Gaussian and binomial outcomes
\item Grid search optimization for threshold locations
\item Cluster-robust standard errors for correlated data
\item Visualization of model fits and log-likelihood surfaces
}
}
\author{
\strong{Maintainer}: Ngo Cheung \email{cheungngo@gmail.com}

Authors:
\itemize{
  \item Kenneth C.Y. Wong \email{mail@cywong.hk}
  \item Hon Cheong So \email{hcso@cuhk.edu.hk}
}

}
